refactor to reflect feedback
issue #3501
