+++ spring-integration-graphql/src/test/java/org/springframework/integration/graphql/outbound/GraphQlMessageHandlerTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2022 the original author or authors.
+ * Copyright 2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,7 +20,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.Duration;
 import java.util.Collections;
-import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
 import java.util.UUID;
@@ -33,7 +32,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.graphql.GraphQlService;
 import org.springframework.graphql.RequestInput;
-import org.springframework.graphql.RequestOutput;
 import org.springframework.graphql.data.method.annotation.Argument;
 import org.springframework.graphql.data.method.annotation.MutationMapping;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
@@ -56,9 +54,9 @@ import org.springframework.stereotype.Controller;
 import org.springframework.stereotype.Repository;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
-import org.springframework.util.AlternativeJdkIdGenerator;
-import org.springframework.util.IdGenerator;
 
+import graphql.ExecutionResult;
+import graphql.ExecutionResultImpl;
 import graphql.execution.reactive.SubscriptionPublisher;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
@@ -67,10 +65,10 @@ import reactor.test.StepVerifier;
 /**
  *
  * @author Daniel Frey
- * @since 6.0
+ *
  */
-@SpringJUnitConfig
-@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
+@SpringJUnitConfig(GraphQlMessageHandlerTests.TestConfig.class)
+@DirtiesContext
 public class GraphQlMessageHandlerTests {
 
 	@Autowired
@@ -83,41 +81,30 @@ public class GraphQlMessageHandlerTests {
 	private PollableChannel errorChannel;
 
 	@Autowired
-	private GraphQlMessageHandler graphQlMessageHandler;
-
-	@Autowired
 	private UpdateRepository updateRepository;
-
-	@Autowired
-	private IdGenerator idGenerator;
-
 
 	@Test
 	@SuppressWarnings("unchecked")
-	void testHandleMessageForQueryWithRequestInputProvided() {
-
-		Locale locale = Locale.getDefault();
-		String executionId = this.idGenerator.generateId().toString();
-		this.graphQlMessageHandler.setLocale(locale);
-		this.graphQlMessageHandler.setExecutionId(executionId);
+	void testHandleMessageForQuery() {
 
 		StepVerifier verifier = StepVerifier.create(
 				Flux.from(this.resultChannel)
 						.map(Message::getPayload)
-						.cast(RequestOutput.class)
+						.cast(ExecutionResult.class)
 				)
 				.consumeNextWith(result -> {
-					assertThat(result).isInstanceOf(RequestOutput.class);
+					assertThat(result).isInstanceOf(ExecutionResultImpl.class);
 					Map<String, Object> data = result.getData();
 					Map<String, Object> testQuery = (Map<String, Object>) data.get("testQuery");
 					assertThat(testQuery.get("id")).isEqualTo("test-data");
-				})
+				}
+				)
 				.thenCancel()
 				.verifyLater();
 
 		this.inputChannel.send(
 				MessageBuilder
-						.withPayload(new RequestInput("{ testQuery { id } }", null, Collections.emptyMap(), locale, executionId))
+						.withPayload(new RequestInput("{ testQuery { id } }", null, Collections.emptyMap()))
 						.build()
 		);
 
@@ -126,33 +113,7 @@ public class GraphQlMessageHandlerTests {
 
 	@Test
 	@SuppressWarnings("unchecked")
-	void testHandleMessageForQueryWithQueryProvided() {
-
-		String executionId = this.idGenerator.generateId().toString();
-		this.graphQlMessageHandler.setExecutionId(executionId);
-
-		String fakeQuery = "{ testQuery { id } }";
-		this.graphQlMessageHandler.setQuery(fakeQuery);
-
-		StepVerifier.create(
-				Mono.from((Mono<RequestOutput>) this.graphQlMessageHandler.handleRequestMessage(MessageBuilder.withPayload(fakeQuery).build()))
-		)
-				.consumeNextWith(result -> {
-					assertThat(result).isInstanceOf(RequestOutput.class);
-					Map<String, Object> data = result.getData();
-					Map<String, Object> testQuery = (Map<String, Object>) data.get("testQuery");
-					assertThat(testQuery.get("id")).isEqualTo("test-data");
-				})
-				.expectComplete()
-				.verify();
-	}
-
-	@Test
-	@SuppressWarnings("unchecked")
-	void testHandleMessageForMutationWithRequestInputProvided() {
-
-		String executionId = this.idGenerator.generateId().toString();
-		this.graphQlMessageHandler.setExecutionId(executionId);
+	void testHandleMessageForMutation() {
 
 		String fakeId = UUID.randomUUID().toString();
 		Update expected = new Update(fakeId);
@@ -160,10 +121,10 @@ public class GraphQlMessageHandlerTests {
 		StepVerifier verifier = StepVerifier.create(
 						Flux.from(this.resultChannel)
 								.map(Message::getPayload)
-								.cast(RequestOutput.class)
+								.cast(ExecutionResult.class)
 				)
 				.consumeNextWith(result -> {
-							assertThat(result).isInstanceOf(RequestOutput.class);
+							assertThat(result).isInstanceOf(ExecutionResultImpl.class);
 							Map<String, Object> data = result.getData();
 							Map<String, Object> update = (Map<String, Object>) data.get("update");
 							assertThat(update.get("id")).isEqualTo(fakeId);
@@ -176,7 +137,7 @@ public class GraphQlMessageHandlerTests {
 
 		this.inputChannel.send(
 				MessageBuilder
-						.withPayload(new RequestInput("mutation { update(id: \"" + fakeId + "\") { id } }", null, Collections.emptyMap(), null, executionId))
+						.withPayload(new RequestInput("mutation { update(id: \"" + fakeId + "\") { id } }", null, Collections.emptyMap()))
 						.build()
 		);
 
@@ -191,22 +152,19 @@ public class GraphQlMessageHandlerTests {
 
 	@Test
 	@SuppressWarnings("unchecked")
-	void testHandleMessageForSubscriptionWithRequestInputProvided() {
-
-		String executionId = this.idGenerator.generateId().toString();
-		this.graphQlMessageHandler.setExecutionId(executionId);
+	void testHandleMessageForSubscription() {
 
 		StepVerifier verifier = StepVerifier.create(
 						Flux.from(this.resultChannel)
 								.map(Message::getPayload)
-								.cast(RequestOutput.class)
-								.mapNotNull(RequestOutput::getData)
+								.cast(ExecutionResult.class)
+								.map(ExecutionResult::getData)
 								.cast(SubscriptionPublisher.class)
 								.map(Flux::from)
 								.flatMap(data -> data)
 				)
-				.consumeNextWith(requestOutput -> {
-					Map<String, Object> results = (Map<String, Object>) requestOutput.getData();
+				.consumeNextWith(executionResult -> {
+					Map<String, Object> results = (Map<String, Object>) executionResult.getData();
 					assertThat(results).containsKey("results");
 
 					Map<String, Object> queryResult = (Map<String, Object>) results.get("results");
@@ -221,7 +179,7 @@ public class GraphQlMessageHandlerTests {
 
 		this.inputChannel.send(
 				MessageBuilder
-						.withPayload(new RequestInput("subscription { results { id } }", null, Collections.emptyMap(), null, executionId))
+						.withPayload(new RequestInput("subscription { results { id } }", null, Collections.emptyMap()))
 						.build()
 		);
 
@@ -233,7 +191,7 @@ public class GraphQlMessageHandlerTests {
 
 		this.inputChannel.send(
 				MessageBuilder
-						.withPayload(new Object())
+						.withPayload("{ testQuery { id } }")
 						.build()
 		);
 
@@ -244,16 +202,18 @@ public class GraphQlMessageHandlerTests {
 				.isInstanceOf(MessageHandlingException.class)
 				.satisfies((ex) -> assertThat((Exception) ex)
 						.hasMessageContaining(
-								"'queryExpression' must not be null"));
+								"Message payload needs to be 'org.springframework.graphql.RequestInput'"));
 
 	}
 
 	@Test
 	void testHandleMessageForMutationWithInvalidPayload() {
 
+		String fakeId = UUID.randomUUID().toString();
+
 		this.inputChannel.send(
 				MessageBuilder
-						.withPayload(new Object())
+						.withPayload("mutation { update(id: \"" + fakeId + "\") { id } }")
 						.build()
 		);
 
@@ -264,7 +224,7 @@ public class GraphQlMessageHandlerTests {
 				.isInstanceOf(MessageHandlingException.class)
 				.satisfies((ex) -> assertThat((Exception) ex)
 						.hasMessageContaining(
-								"'queryExpression' must not be null"));
+								"Message payload needs to be 'org.springframework.graphql.RequestInput'"));
 
 	}
 
@@ -273,7 +233,7 @@ public class GraphQlMessageHandlerTests {
 
 		this.inputChannel.send(
 				MessageBuilder
-						.withPayload(new Object())
+						.withPayload("subscription { results { id } }")
 						.build()
 		);
 
@@ -284,7 +244,7 @@ public class GraphQlMessageHandlerTests {
 				.isInstanceOf(MessageHandlingException.class)
 				.satisfies((ex) -> assertThat((Exception) ex)
 						.hasMessageContaining(
-								"'queryExpression' must not be null"));
+								"Message payload needs to be 'org.springframework.graphql.RequestInput'"));
 
 	}
 
@@ -299,11 +259,6 @@ public class GraphQlMessageHandlerTests {
 
 		@QueryMapping
 		public Mono<QueryResult> testQuery() {
-			return Mono.just(new QueryResult("test-data"));
-		}
-
-		@QueryMapping
-		public Mono<QueryResult> testQueryById(@Argument String id) {
 			return Mono.just(new QueryResult("test-data"));
 		}
 
@@ -401,12 +356,6 @@ public class GraphQlMessageHandlerTests {
 		AnnotatedControllerConfigurer annotatedDataFetcherConfigurer() {
 
 			return new AnnotatedControllerConfigurer();
-		}
-
-		@Bean
-		IdGenerator idGenerator() {
-
-			return new AlternativeJdkIdGenerator();
 		}
 
 	}
